# dist/openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "alarm-manager API",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "[pdjr-skplugin-alarm-manager](https://github.com/pdjr-signalk/pdjr-skplugin-alarm-manager)\nis a plugin for the\n[Signal K node server](https://github.com/SignalK/signalk-server).\n\nA Signal K server hosting the plugin will present this API under\nthe root https://*server_ip*:*server_port*/plugins/alarm-manager.\n"
  },
  "externalDocs": {
    "description": "Plugin README document",
    "url": "https://github.com/pdjr-signalk/pdjr-skplugin-alarm-manager#readme"
  },
  "paths": {
    "/keys": {
      "get": {
        "description": "Get a digest of all Signal K keys which are being monitored for\nalarm activity (i.e. all keys where the associated metadata\nincludes some zone definitions).\n",
        "responses": {
          "200": {
            "description": "Success.\nThe response body is an array containing the keys of all\ncurrently monitored data items.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error.\nThe server was unable to service the request.       \n"
          }
        }
      }
    },
    "/digest": {
      "get": {
        "description": "Get a digest of all active alarm notifications.\n",
        "responses": {
          "200": {
            "description": "Success.\nThe response body is an object with the paths of alarm\nnotifications as property names and property values which\ninclude notification and suppression details.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "notification": {
                          "properties": {
                            "state": {
                              "type": "string",
                              "description": "Alarm state of the notification."
                            },
                            "method": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "message": {
                              "type": "string"
                            },
                            "id": {
                              "type": "string"
                            }
                          }
                        },
                        "suppressedOutputs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error.\nThe server was unable to service the request.\n"
          }
        }
      }
    },
    "/outputs": {
      "get": {
        "description": "Get the state of all configured outputs.\n",
        "responses": {
          "200": {
            "description": "Success.\nThe response body is an object with the names of alarm\noutput channels as property names and the current channel\nstate a the property value.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "additionalProperties": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error.\nThe server was unable to service the request.\n"
          }
        }
      }
    },
    "/output/{output}": {
      "get": {
        "description": "Get the current state of the output specified by {name}.\n",
        "parameters": [
          {
            "output": null,
            "name": "output",
            "in": "path",
            "description": "Name of the output channel whose state is required.\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.\nThe response body contains the value 1 to indicate that\nthe output is ON, or 0 to indicate that the output is OFF,\nor -1 to indicate that the output state is undefined (i.e.\nhas not yet been set).\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number",
                  "minimum": -1,
                  "maximum": 1
                }
              }
            }
          },
          "404": {
            "description": "Error.\nThe request was invalid (bad or missing output channel name).\n"
          },
          "500": {
            "description": "Error.\nThe server was unable to service the request.\n"
          }
        }
      }
    },
    "/suppress/{output}": {
      "patch": {
        "description": "Suppress output on output channel {name}.\n",
        "parameters": [
          {
            "output": null,
            "name": "output",
            "in": "path",
            "description": "Identifier of the output channel which should be suppressed.\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.\nOutput channel has been suppressed.\n"
          },
          "404": {
            "description": "Error.\nThe request was invalid (bad or missing output channel name).\n"
          },
          "500": {
            "description": "Error.\nThe server was unable to service the request.\n"
          }
        }
      }
    }
  }
}